feature(S0MT-24): 사용자 서비스 impl구현체 구조로 변경
